Added getEmptyInputFormatClassName
