+++ format-common/src/main/java/io/cdap/plugin/format/input/EmptyInputFormat.java
@@ -15,60 +15,12 @@
  */
 
 package io.cdap.plugin.format.input;
-import org.apache.hadoop.mapreduce.InputFormat;
-import org.apache.hadoop.mapreduce.InputSplit;
-import org.apache.hadoop.mapreduce.JobContext;
-import org.apache.hadoop.mapreduce.RecordReader;
-import org.apache.hadoop.mapreduce.TaskAttemptContext;
-
-import java.io.IOException;
-import java.util.Collections;
-import java.util.List;
 
 /**
  * An InputFormat that returns no data.
  * @param <K> the key class
  * @param <V> the value class
  */
-public class EmptyInputFormat<K, V> extends InputFormat<K, V> {
-
-  @Override
-  public List<InputSplit> getSplits(JobContext context) throws IOException, InterruptedException {
-    return Collections.emptyList();
-  }
-
-  @Override
-  public RecordReader<K, V> createRecordReader(InputSplit split, TaskAttemptContext context) {
-    return new RecordReader<K, V>() {
-      @Override
-      public void initialize(InputSplit split, TaskAttemptContext context) {
-        // do nothing
-      }
-
-      @Override
-      public boolean nextKeyValue() {
-        return false;
-      }
-
-      @Override
-      public K getCurrentKey() {
-        return null;
-      }
-
-      @Override
-      public V getCurrentValue() {
-        return null;
-      }
-
-      @Override
-      public float getProgress() {
-        return 1.0F;
-      }
-
-      @Override
-      public void close() {
-        // nothing to do
-      }
-    };
-  }
+public class EmptyInputFormat<K, V> extends AbstractEmptyInputFormat<K, V> {
+  // no-op
 }
